falta de validaciones, correccion de errores, bonito y documento

+++ src/main/java/com/eep/Coches/Entity/Marca.java
@@ -8,25 +8,25 @@ public class Marca {
     @Id
     @GeneratedValue
     @Column(name = "id")
-    private Long id;
+    private int id;
     @Column(name = "marca")
     String marca;
 
     public Marca() {
     }
-    public Marca(Long id, String marca) {
+    public Marca(int id, String marca) {
         this.id = id;
         this.marca = marca;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
     public String getMarca() {
         return marca;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
     public void setMarca(String marca) {
